requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# AI & ML
sentence-transformers==2.3.1
faiss-cpu==1.7.4
groq==0.4.1
langchain==0.1.0
langchain-community==0.0.10

# Document Processing
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0

# Database
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0